fix: extract DAYS constants from 'use client' to fix SSG build
